+++ Assets/Packages/DraasGames.Core/Runtime/Infrastructure/Installers/ResourcesViewInstaller.cs
@@ -1,7 +1,9 @@
-﻿using DraasGames.Core.Runtime.UI.PresenterNavigationService.Concrete;
+﻿using DraasGames.Core.Runtime.Infrastructure.Extensions;
+using DraasGames.Core.Runtime.UI.PresenterNavigationService.Concrete;
 using DraasGames.Core.Runtime.UI.Views.Concrete;
 using DraasGames.Core.Runtime.UI.Views.Concrete.ViewContainers;
 using DraasGames.Core.Runtime.UI.Views.Concrete.ViewProviders;
+using Sirenix.OdinInspector;
 using UnityEngine;
 using Zenject;
 
@@ -9,26 +11,33 @@ namespace DraasGames.Core.Runtime.Infrastructure.Installers
 {
     public class ResourcesViewInstaller : MonoInstaller
     {
-        [SerializeField] 
+        [SerializeField, Required, AssetsOnly] 
         private ResourcesViewContainer _resourcesViewContainer;
 
+        [SerializeField]
+        [InfoBox("Installer should not be placed on ProjectContext if MoveIntoAllSubcontainers is disabled" +
+                 " or placed at scene context if MoveIntoAllSubcontainers is enabled!",
+            InfoMessageType.Error,
+            VisibleIf = nameof(ShouldShowProjectContextError))]
+        private bool _moveIntoAllSubContainers = false;
+        
         public override void InstallBindings()
         {
             Container
                 .BindInterfacesAndSelfTo<ViewFactory>()
                 .FromNew()
                 .AsSingle()
-                .MoveIntoAllSubContainers();;
+                .MoveIntoAllSubContainersConditional(_moveIntoAllSubContainers);
             
             Container
                 .BindInterfacesAndSelfTo<ViewRouter>()
                 .AsSingle()
-                .MoveIntoAllSubContainers();
+                .MoveIntoAllSubContainersConditional(_moveIntoAllSubContainers);
 
             Container
                 .BindInterfacesTo<PresenterNavigationService>()
                 .AsSingle()
-                .MoveIntoAllSubContainers();
+                .MoveIntoAllSubContainersConditional(_moveIntoAllSubContainers);
             
             BindViewRetrieval();
         }
@@ -38,12 +47,27 @@ namespace DraasGames.Core.Runtime.Infrastructure.Installers
             Container
                 .Bind<ResourcesViewContainer>()
                 .FromInstance(_resourcesViewContainer)
-                .AsSingle().MoveIntoAllSubContainers();
+                .AsSingle()
+                .MoveIntoAllSubContainersConditional(_moveIntoAllSubContainers);
                 
             Container
                 .BindInterfacesTo<ResourcesViewProviderAsync>()
                 .FromNew()
-                .AsSingle().MoveIntoAllSubContainers();
+                .AsSingle()
+                .MoveIntoAllSubContainersConditional(_moveIntoAllSubContainers);
         }
+        
+#if UNITY_EDITOR
+        // TODO to general validation method
+        private bool ShouldShowProjectContextError()
+        {
+            return !IsProjectContext() && _moveIntoAllSubContainers;
+        }
+
+        private bool IsProjectContext()
+        {
+            return gameObject.name == "ProjectContext" || transform.root.name == "ProjectContext";
+        }
+#endif
     }
 }