+++ Assets/Packages/DraasGames.Core/Runtime/Infrastructure/Loaders/Abstract/IAssetLoader.cs
@@ -7,14 +7,46 @@ namespace DraasGames.Core.Runtime.Infrastructure.Loaders.Abstract
 {
     public interface IAssetLoader
     {
-        public UniTask<T> LoadAsync<T>(AssetReference assetReference, ILifetime lifetime, Action<float> onLoadProgress = null);
+        /// <summary>
+        /// Load asset by AssetReference
+        /// </summary>
+        /// <param name="assetReference"></param>
+        /// <param name="lifetime">Asset lifetime. If lifetime is null, it will use ProjectLifeTime and will never be unloaded</param>
+        /// <param name="onLoadProgress"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public UniTask<T> LoadAsync<T>(AssetReference assetReference, ILifetime lifetime = null, Action<float> onLoadProgress = null);
 
-        public UniTask<T> LoadAsync<T>(string key, ILifetime lifetime, Action<float> onLoadProgress = null);
+        /// <summary>
+        /// Load asset by key
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="lifetime">Asset lifetime. If lifetime is null, it will use ProjectLifeTime and will never be unloaded</param>
+        /// <param name="onLoadProgress"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public UniTask<T> LoadAsync<T>(string key, ILifetime lifetime = null, Action<float> onLoadProgress = null);
 
-        public UniTask<T> LoadWithComponentAsync<T>(AssetReference assetReference, ILifetime lifetime, Action<float> onLoadProgress = null)
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="assetReference"></param>
+        /// <param name="lifetime">Asset lifetime. If lifetime is null, it will use ProjectLifeTime and will never be unloaded</param>
+        /// <param name="onLoadProgress"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public UniTask<T> LoadWithComponentAsync<T>(AssetReference assetReference, ILifetime lifetime = null, Action<float> onLoadProgress = null)
             where T : Component;
 
-        public UniTask<T> LoadWithComponentAsync<T>(string key, ILifetime lifetime, Action<float> onLoadProgress = null) 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="lifetime">Asset lifetime. If lifetime is null, it will use ProjectLifeTime and will never be unloaded</param>
+        /// <param name="onLoadProgress"></param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        public UniTask<T> LoadWithComponentAsync<T>(string key, ILifetime lifetime = null, Action<float> onLoadProgress = null) 
             where T : Component;
     }
 }