+++ Assets/Packages/DraasGames.Core/Runtime/Infrastructure/Loaders/Abstract/ISceneLifetimeProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace DraasGames.Core.Runtime.Infrastructure.Loaders.Abstract
-{
-    public interface ISceneLifetimeProvider
-    {
-        public ILifetime Lifetime { get; }
-    }
-}