+++ Assets/Packages/DraasGames.Core/Runtime/Infrastructure/Loaders/Concrete/ScopeLifetimeProvider.cs
@@ -3,13 +3,13 @@ using DraasGames.Core.Runtime.Infrastructure.Loaders.Abstract;
 
 namespace DraasGames.Core.Runtime.Infrastructure.Loaders.Concrete
 {
-    public class ProjectLifetimeProvider : IProjectLifetimeProvider, IDisposable
+    public class ScopeLifetimeProvider : IScopeLifetimeProvider, IDisposable
     {
-        public ILifetime Lifetime { get; } = new Lifetime();
+        public ILifetime ScopeLifetime { get; } = new Lifetime();
 
         public void Dispose()
         {
-            Lifetime?.Dispose();
+            ScopeLifetime?.Dispose();
         }
     }
 }